hf mf chk refactoring
