+++ armsrc/mifarecmd.h
@@ -17,7 +17,6 @@
 #include "apps.h"
 #include "util.h"
 #include "string.h"
-#include "usb_cdc.h"
 
 #include "iso14443crc.h"
 #include "iso14443a.h"
